refactor: move nextauth endpoint to app
