+++ src/lib/session.ts
@@ -1,4 +1,4 @@
-import { getServerSession } from 'next-auth/next'
+import { getServerSession } from 'next-auth'
 
 import { authOptions } from '@/lib/auth'
