--- src/pages/api/auth/[...nextauth].ts
@@ -1,7 +1,0 @@
-import NextAuth from 'next-auth'
-
-import { authOptions } from '@/lib/auth'
-
-export { authOptions }
-
-export default NextAuth(authOptions)
