+++ src/pages/api/bookmarks/task/[taskId].ts
@@ -3,6 +3,7 @@ import type { NextApiRequest, NextApiResponse } from 'next'
 import { unstable_getServerSession } from 'next-auth'
 
 import { BookmarkCDSchema } from '@/lib/api/schema/bookmark'
+import { authOptions } from '@/lib/auth'
 import prisma from '@/lib/prisma'
 import {
   methodNotAllowed,
@@ -10,8 +11,6 @@ import {
   ok,
   badRequest
 } from '@/utils/response'
-
-import { authOptions } from '../../auth/[...nextauth]'
 
 export default async function handler(
   req: NextApiRequest,
