+++ src/pages/api/submissions/score.ts
@@ -4,10 +4,9 @@ import { Prisma } from '@prisma/client'
 import { unstable_getServerSession } from 'next-auth'
 
 import { getAllTaskForUser } from '@/lib/api/queries/getAllTaskForUser'
+import { authOptions } from '@/lib/auth'
 import prisma from '@/lib/prisma'
 import { methodNotAllowed, ok, unauthorized } from '@/utils/response'
-
-import { authOptions } from '../auth/[...nextauth]'
 
 export default async function handler(
   req: NextApiRequest,
