+++ src/pages/api/tasks/[id]/index.ts
@@ -4,6 +4,7 @@ import { unstable_getServerSession } from 'next-auth'
 
 import checkUserPermissionOnTask from '@/lib/api/queries/checkUserPermissionOnTask'
 import { IndividualTaskSchema, TaskSchema } from '@/lib/api/schema/tasks'
+import { authOptions } from '@/lib/auth'
 import prisma from '@/lib/prisma'
 import {
   unauthorized,
@@ -12,8 +13,6 @@ import {
   forbidden,
   badRequest
 } from '@/utils/response'
-
-import { authOptions } from '../../auth/[...nextauth]'
 
 export default async function handler(
   req: NextApiRequest,
